fix: rename firestoreRepository to artistsRepository for clarity
