+++ src/api/v1/repositories/artistsRepository.ts
